backend/backend-func/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
flask
flask-cors
werkzeug
# ...and any other dependencies from your main backend/requirements.txt